Move the fallback Numpy dtype conversion to the `numpy` mapping
